+++ pylace/src/lib.rs
@@ -1019,6 +1019,22 @@ impl CoreEngine {
         Ok(val.into())
     }
 
+    #[pyo3(signature=(target, given=None, state_ixs=None))]
+    fn mean(
+        &self,
+        target: &Bound<PyAny>,
+        given: Option<&Bound<PyDict>>,
+        state_ixs: Option<Vec<usize>>,
+    ) -> PyResult<Option<f64>> {
+        let col_ix = value_to_index(target, &self.col_indexer)?;
+        let given = dict_to_given(given, &self.engine, &self.col_indexer)?;
+        let val = self
+            .engine
+            .mean(col_ix, &given, state_ixs.as_deref())
+            .map_err(|err| PyErr::new::<PyValueError, _>(format!("{err}")))?;
+        Ok(val.into())
+    }
+
     /// Forward the Markov chains
     ///
     /// Parameters